Вывод export.txt перенесен из compiler.c в codegen.c
